+++ insert/src/main/java/it/gas/altichierock/insert/ReceiptKeeper.java
@@ -12,7 +12,7 @@ public class ReceiptKeeper {
 		model = new ReceiptTableModel(this);
 	}
 	
-	public void add(int qty, String descr, float price) {
+	public void add(int qty, String descr, double price) {
 		Row r = new Row();
 		r.quantity = qty;
 		r.description = descr;
@@ -43,11 +43,11 @@ public class ReceiptKeeper {
 		model.fireTableDataChanged();
 	}
 	
-	public float getPrice(int row) {
+	public double getPrice(int row) {
 		return list.get(row).price;
 	}
 	
-	public void setPrice(int row, float price) {
+	public void setPrice(int row, double price) {
 		list.get(row).price = price;
 		model.fireTableDataChanged();
 	}
@@ -64,6 +64,6 @@ public class ReceiptKeeper {
 	private class Row {
 		private int quantity;
 		private String description;
-		private float price;
+		private double price;
 	}
 }
